nrf_usbd

Struct Usbd

pub struct Usbd<T: UsbPeripheral> { /* private fields */ }

USB device implementation.

This type implements the UsbBus trait and can be passed to a UsbBusAllocator to configure and use the USB device.

Implementations

impl<T: UsbPeripheral> Usbd<T>

pub fn new(periph: T) -> Self

Creates a new USB device wrapper, taking ownership of the raw peripheral.

Parameters

periph: The raw USBD peripheral.

pub fn device_address(&self) -> u8

Fetches the address assigned to the device (only valid when device is configured).